chore: moved assets-generation to it's own package
