+++ .github/workflows/ci.yml
@@ -57,7 +57,9 @@ jobs:
           cache: 'pnpm'
 
       - name: Install dependencies
-        run: pnpm install
+        # The docs-asset harness (ffmpeg + a second Playwright) is never run in
+        # CI, so exclude it — that's what keeps the ffmpeg binary out of CI.
+        run: pnpm install --filter "!@humblehire/docs-assets"
 
       - name: Install Playwright browsers
         run: pnpm exec playwright install --with-deps chromium
@@ -92,7 +94,9 @@ jobs:
           cache: 'pnpm'
 
       - name: Install dependencies
-        run: pnpm install
+        # The docs-asset harness (ffmpeg + a second Playwright) is never run in
+        # CI, so exclude it — that's what keeps the ffmpeg binary out of CI.
+        run: pnpm install --filter "!@humblehire/docs-assets"
 
       - name: Build
         env:
@@ -121,7 +125,9 @@ jobs:
           cache: 'pnpm'
 
       - name: Install dependencies
-        run: pnpm install
+        # The docs-asset harness (ffmpeg + a second Playwright) is never run in
+        # CI, so exclude it — that's what keeps the ffmpeg binary out of CI.
+        run: pnpm install --filter "!@humblehire/docs-assets"
 
       - name: Install Playwright browsers
         run: pnpm exec playwright install --with-deps chromium
@@ -158,7 +164,9 @@ jobs:
           cache: 'pnpm'
 
       - name: Install dependencies
-        run: pnpm install
+        # The docs-asset harness (ffmpeg + a second Playwright) is never run in
+        # CI, so exclude it — that's what keeps the ffmpeg binary out of CI.
+        run: pnpm install --filter "!@humblehire/docs-assets"
 
       - name: Build Storybook
         run: pnpm build-storybook
