--- assets/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"extends": "../.svelte-kit/tsconfig.json",
-	"include": ["./**/*.ts", "../.svelte-kit/ambient.d.ts", "../src/app.d.ts"]
-}
