+++ e2e/editor.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { useBridge } from './helpers/bridge';
+import { useBridge } from '@humblehire/test-bridge';
 import { fillInlineField, gotoHome, waitForAutosave, reorder } from './helpers/ui';
 import type { CV } from '$lib/types/cv';
