+++ package.json
@@ -18,12 +18,13 @@
 		"storybook": "storybook dev -p 6006",
 		"build-storybook": "storybook build",
 		"generate-pwa-assets": "pwa-assets-generator",
-		"assets:generate": "playwright test -c playwright.assets.config.ts"
+		"assets:generate": "pnpm --filter @humblehire/docs-assets generate"
 	},
 	"devDependencies": {
 		"@chromatic-com/storybook": "^5.0.1",
 		"@eslint/compat": "^1.2.5",
 		"@eslint/js": "^9.18.0",
+		"@humblehire/test-bridge": "workspace:*",
 		"@internationalized/date": "^3.12.0",
 		"@lucide/svelte": "^1.16.0",
 		"@playwright/test": "^1.49.1",
@@ -49,13 +50,11 @@
 		"eslint-config-prettier": "^10.0.1",
 		"eslint-plugin-storybook": "^10.3.0",
 		"eslint-plugin-svelte": "^3.0.0",
-		"ffmpeg-static": "^5.3.0",
 		"globals": "^16.0.0",
 		"husky": "^9.1.7",
 		"jsdom": "^26.0.0",
 		"lint-staged": "^17.0.7",
 		"mode-watcher": "^1.1.0",
-		"mouse-helper": "^1.0.7",
 		"playwright": "^1.58.2",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
