+++ test-bridge/package.json
@@ -1,0 +1,10 @@
+{
+	"name": "@humblehire/test-bridge",
+	"version": "0.0.0",
+	"private": true,
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts"
+	},
+	"types": "./src/index.ts"
+}
